Fix 'unrecognized command' error after out of sequence command
Various commands check first whether it is the appropriate time in the
flow of the normal smtp sequence for the user to issue that command,
and if so immediately reply with a 503 error without consuming the rest
of the user input on the line (i.e. any args to the command and the \r\n)

When this occurs, the leftover args in the buffer are then processed as
the start of another command usually resulting in an 'unrecognized command'
error message.

This patch fixes them by simply moving the call to `read_line` which would
immediately follow the sequence check to before it so that the rest of
the line is consumed regardless.

The only change in behavior as a result of this patch is that if a user
sends an overly long line of input starting with a command that is also
inappropriate for this moment in the transaction, they will receive the
500 error for the line being too long first instead of the 503 out of
sequence error, and they will not receive a second buggy 500 unrecognized
command error which would likely cause a breakdown in the normal server
client lockstep process and potentially cause many knock on issues.

Signed-off-by: Charles Mirabile <cmirabil@redhat.com>